haystack-api/requirements.txt
fastapi
farm-haystack
uvicorn